fix(libultraship): back-port Kenix3 #1126 warning fixes for soh /WX
soh@develop builds with warnings-as-errors (#6733) and trips on two LUS headers
we vendor (lus_gbi.h GIMMCMD redefinition, BitConverter.h size_t->int conversion).
soh tracks LUS port-maintenance (has #1126); we vendor LUS main (doesn't yet), so
back-port just #1126 additively. Did NOT switch to port-maintenance (carries #1103,
intentionally skipped). Documented in docs/merges/2026-06-20.md.

+++ libultraship/include/ship/utils/binarytools/BitConverter.h
@@ -241,12 +241,14 @@ class BitConverter {
         switch (firstByte) {
             case 0x37: // v64
                 for (size_t pos = 0; pos < (romSize / 2); pos++) {
-                    ((uint16_t*)rom)[pos] = ToUInt16BE(rom, pos * 2);
+                    // ComboShip: back-port of Kenix3 libultraship #1126 (Fix warnings) — static_cast.
+                    ((uint16_t*)rom)[pos] = ToUInt16BE(rom, static_cast<int32_t>(pos * 2));
                 }
                 break;
             case 0x40: // n64
                 for (size_t pos = 0; pos < (romSize / 4); pos++) {
-                    ((uint32_t*)rom)[pos] = ToUInt32BE(rom, pos * 4);
+                    // ComboShip: back-port of Kenix3 libultraship #1126 (Fix warnings) — static_cast.
+                    ((uint32_t*)rom)[pos] = ToUInt32BE(rom, static_cast<int32_t>(pos * 4));
                 }
                 break;
             case 0x80: // z64
